infra/scripts/agent_scripts/requirements.txt
azure-identity==1.23.0
azure-ai-projects==1.0.0b12
